Comment-out IterableLike Scala specs, remove IterableFutureAssertionsSpec
